// src/types/dashboard/sigunguEventWithDates.ts
export type SigunguEventWithDates = {
  contentTypeId: string;
  contentId: string;
  title: string;
  startDate: string;
  endDate: string;
  zipCode: string;
  address: string;
  latitude: number;
  longitude: number;
  telephone: string;
  originalImageUrl: string;
  thumbnailImageUrl: string;
};